+++ angular_forms/lib/src/directives/checkbox_value_accessor.dart
@@ -42,5 +42,7 @@ class CheckboxControlValueAccessor extends Object
   }
 
   @override
-  void onDisabledChanged(bool isDisabled) {}
+  void onDisabledChanged(bool isDisabled) {
+    _element.disabled = isDisabled;
+  }
 }