+++ angular_forms/lib/src/directives/default_value_accessor.dart
@@ -2,6 +2,8 @@ import 'dart:html';
 import 'dart:js_util' as js_util;
 
 import 'package:angular/angular.dart';
+import 'package:angular_forms/src/directives/shared.dart'
+    show setElementDisabled;
 
 import 'control_value_accessor.dart';
 
@@ -43,5 +45,7 @@ class DefaultValueAccessor extends Object
   }
 
   @override
-  void onDisabledChanged(bool isDisabled) {}
+  void onDisabledChanged(bool isDisabled) {
+    setElementDisabled(_element, isDisabled);
+  }
 }