+++ angular_forms/lib/src/directives/number_value_accessor.dart
@@ -41,5 +41,7 @@ class NumberValueAccessor extends Object
   }
 
   @override
-  void onDisabledChanged(bool isDisabled) {}
+  void onDisabledChanged(bool isDisabled) {
+    _element.disabled = isDisabled;
+  }
 }